Setup/Cleanup ACI on windows

+++ cleanup.ps1
@@ -2,4 +2,7 @@ param (
     [string]$OracleContainerName
 )
 
-$ignore = az container delete --resource-group GitHubActions-RG --name $OracleContainerName --yes
+if ($Env:RUNNER_OS -eq "Windows") 
+{
+    $ignore = az container delete --resource-group GitHubActions-RG --name $OracleContainerName --yes
+}
